+++ Changelog.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Introduce the algorithm `proximal_gradient_method` along with `ManifoldProximalGradientObjective`, `ProximalGradientMethodState`, `ProxGradAcceleration`.
+* Introduce the algorithm `proximal_gradient_method` along with `ManifoldProximalGradientObjective`, `ProximalGradientMethodState`, `ProximalGradientMethodAcceleration`.
 * Add `ProximalGradientMethodBacktracking` stepsize.
 * Add `StopWhenGradientMappingNormLess` stopping criterion.
 * Added back functions `estimate_sectional_curvature`, `ζ_1`, `ζ_2`, `close_point` from `convex_bundle_method`; the function call can stay the same as before since there is a curvature estimation fallback